requirements.txt
colorzero==2.0
gpiozero==2.0.1
pillow==11.3.0
setuptools==80.9.0
spidev==3.7
